Refactored web server port assigning from app class to config class.
